Выполняет разделение на модули
