+++ index.html
@@ -228,7 +228,7 @@
   </template>
 
   <!-- Скрипты -->
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
   <script src="js/functions.js"></script>
 
 </body>
