+++ js/data.js
@@ -1,0 +1,36 @@
+//Задаём количество идентификаторов
+const idCount = 25;
+
+//Задаём массивы с именами, комментариями и описаниями
+const NAMES = [
+  'Денетор',
+  'Элронд',
+  'Исильдур',
+  'Теоден',
+  'Арвен',
+  'Эовин',
+  'Мириэль',
+  'Анайрэ'
+];
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const DESCRIPTIONS = [
+  'Я на природе',
+  'Горы перед закатом',
+  'Прекрасный восход',
+  'Джонни, они на деревьях!',
+  'Моя прелесть',
+  'На прогулке с Кексом',
+  'В гостях хорошо, а дома лучше',
+  'Только посмотрите на эту прелесть'
+];
+
+export {idCount, NAMES, COMMENTS, DESCRIPTIONS};
